new1
new1

+++ Summer2021Lab3/src/summer2021lab3/Summer2021Lab3.java
@@ -4,8 +4,8 @@ package summer2021lab3;
 public class Summer2021Lab3 {
 
     public static void main(String[] args) {
-        float [] points={1.00f,2.00f,3.00f,4.00f,5.00f};
-        for(float i:points)
+        int[] points={1,2,3,4,5};
+        for(int i:points)
             System.out.println(i);
     }
 }
